autoValue and defaultValue rewrite
