+++ lib/SimpleSchema.js
@@ -842,6 +842,9 @@ function getDefaultAutoValueFunction(defaultValue) {
     // want any fields missing from that object to be added if they have default values
     if (this.operator === '$push') return defaultValue;
 
+    // If parent is set, we should update this position instead of $setOnInsert
+    if (this.parentField().isSet) return defaultValue;
+
     // We don't know whether it's an upsert, but if it's not, this seems to be ignored,
     // so this is a safe way to make sure the default value is added on upsert insert.
     return { $setOnInsert: defaultValue };
