+++ lib/SimpleSchema.tests.js
@@ -220,13 +220,9 @@ describe('SimpleSchema', function () {
     expect(cleaned).toEqual({
       $set: {
         nested: [
-          { somethingOptional: 'q' },
-          { somethingOptional: 'z' },
+          { somethingAutovalued: 'x', somethingOptional: 'q' },
+          { somethingAutovalued: 'x', somethingOptional: 'z' },
         ],
-      },
-      $setOnInsert: {
-        'nested.0.somethingAutovalued': 'x',
-        'nested.1.somethingAutovalued': 'x',
       },
     });
   });
