+++ lib/clean.js
@@ -1,6 +1,6 @@
 import MongoObject from 'mongo-object';
 import isEmpty from 'lodash.isempty';
-import { looksLikeModifier } from './utility.js';
+import { looksLikeModifier } from './utility';
 import { SimpleSchema } from './SimpleSchema';
 import convertToProperType from './clean/convertToProperType';
 import setAutoValues from './clean/setAutoValues';
