+++ lib/clean.tests.js
@@ -488,950 +488,6 @@ describe('clean', function () {
     });
   });
 
-  describe('autoValue', function () {
-    it('has correct information in function context - empty', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(false);
-            expect(this.value).toBe(undefined);
-            expect(this.operator).toBe(null);
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(false);
-            expect(foo.value).toBe(undefined);
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(false);
-            expect(fooSibling.value).toBe(undefined);
-            expect(fooSibling.operator).toBe(null);
-          },
-        },
-      });
-      schema.clean({});
-    });
-
-    it('has correct information in function context - normal other key', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(false);
-            expect(this.value).toBe(undefined);
-            expect(this.operator).toBe(null);
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(true);
-            expect(foo.value).toEqual('clown');
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(true);
-            expect(fooSibling.value).toEqual('clown');
-            expect(fooSibling.operator).toBe(null);
-          },
-        },
-      });
-      schema.clean({
-        foo: 'clown',
-      });
-    });
-
-    it('has correct information in function context - normal self and other key', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(true);
-            expect(this.value).toBe(true);
-            expect(this.operator).toBe(null);
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(true);
-            expect(foo.value).toEqual('clown');
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(true);
-            expect(fooSibling.value).toEqual('clown');
-            expect(fooSibling.operator).toBe(null);
-          },
-        },
-      });
-      schema.clean({
-        foo: 'clown',
-        bar: true,
-      });
-    });
-
-    it('has correct information in function context - normal self and no other key with unset', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(true);
-            expect(this.value).toBe(false);
-            expect(this.operator).toBe(null);
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(false);
-            expect(foo.value).toBe(undefined);
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(false);
-            expect(fooSibling.value).toBe(undefined);
-            expect(fooSibling.operator).toBe(null);
-            this.unset();
-          },
-        },
-      });
-      const doc = {
-        bar: false,
-      };
-      expect(schema.clean(doc)).toEqual({});
-    });
-
-    it('has correct information in function context - $set self and no other key', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(true);
-            expect(this.value).toBe(false);
-            expect(this.operator).toBe('$set');
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(false);
-            expect(foo.value).toBe(undefined);
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(false);
-            expect(fooSibling.value).toBe(undefined);
-            expect(fooSibling.operator).toBe(null);
-          },
-        },
-      });
-
-      const doc = {
-        $set: {
-          bar: false,
-        },
-      };
-      schema.clean(doc);
-      expect(doc).toEqual({
-        $set: {
-          bar: false,
-        },
-      });
-    });
-
-    it('has correct information in function context - $set self and another key and change self', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(true);
-            expect(this.value).toBe(false);
-            expect(this.operator).toEqual('$set');
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(true);
-            expect(foo.value).toEqual('clown');
-            expect(foo.operator).toEqual('$set');
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(true);
-            expect(fooSibling.value).toEqual('clown');
-            expect(fooSibling.operator).toEqual('$set');
-
-            return true;
-          },
-        },
-      });
-
-      const doc = {
-        $set: {
-          foo: 'clown',
-          bar: false,
-        },
-      };
-      expect(schema.clean(doc)).toEqual({
-        $set: {
-          foo: 'clown',
-          bar: true,
-        },
-      });
-    });
-
-    it('has correct information in function context - adds $set when missing', function () {
-      const schema = new SimpleSchema({
-        foo: {
-          type: String,
-          optional: true,
-        },
-        bar: {
-          type: Boolean,
-          optional: true,
-          autoValue() {
-            expect(this.isSet).toBe(false);
-            expect(this.value).toBe(undefined);
-            expect(this.operator).toEqual('$set');
-
-            const foo = this.field('foo');
-            expect(foo.isSet).toBe(false);
-            expect(foo.value).toBe(undefined);
-            expect(foo.operator).toBe(null);
-
-            const fooSibling = this.siblingField('foo');
-            expect(fooSibling.isSet).toBe(false);
-            expect(fooSibling.value).toBe(undefined);
-            expect(fooSibling.operator).toBe(null);
-
-            return {
-              $set: true,
-            };
-          },
-        },
-      });
-
-      expect(schema.clean({}, { isModifier: true })).toEqual({
-        $set: {
-          bar: true,
-        },
-      });
-    });
-
-    it('content autoValues', function () {
-      const schema = new SimpleSchema({
-        content: {
-          type: String,
-          optional: true,
-        },
-        updatesHistory: {
-          type: Array,
-          optional: true,
-          autoValue() {
-            const content = this.field('content');
-            if (content.isSet) {
-              if (!this.operator) {
-                return [{
-                  date: new Date('2017-01-01T00:00:00.000Z'),
-                  content: content.value,
-                }];
-              }
-              return {
-                $push: {
-                  date: new Date('2017-01-01T00:00:00.000Z'),
-                  content: content.value,
-                },
-              };
-            }
-          },
-        },
-        'updatesHistory.$': {
-          type: Object,
-        },
-        'updatesHistory.$.content': {
-          type: String,
-          optional: true,
-        },
-      });
-
-      // Normal object
-      let result = schema.clean({ content: 'foo' });
-      expect(result).toEqual({
-        content: 'foo',
-        updatesHistory: [
-          {
-            date: new Date('2017-01-01T00:00:00.000Z'),
-            content: 'foo',
-          },
-        ],
-      });
-
-      // $set
-      result = schema.clean({ $set: { content: 'foo' } });
-      expect(result).toEqual({
-        $set: {
-          content: 'foo',
-        },
-        $push: {
-          updatesHistory: {
-            date: new Date('2017-01-01T00:00:00.000Z'),
-            content: 'foo',
-          },
-        },
-      });
-    });
-
-    it('array of objects autoValues', function () {
-      const schema = new SimpleSchema({
-        avArrayOfObjects: {
-          type: Array,
-          optional: true,
-        },
-        'avArrayOfObjects.$': {
-          type: Object,
-        },
-        'avArrayOfObjects.$.a': {
-          type: String,
-        },
-        'avArrayOfObjects.$.foo': {
-          type: String,
-          autoValue() {
-            return 'bar';
-          },
-        },
-      });
-
-      let result = schema.clean({
-        $push: {
-          avArrayOfObjects: {
-            a: 'b',
-          },
-        },
-      });
-
-      expect(result).toEqual({
-        $push: {
-          avArrayOfObjects: {
-            a: 'b',
-            foo: 'bar',
-          },
-        },
-      });
-
-      result = schema.clean({
-        $set: {
-          avArrayOfObjects: [{
-            a: 'b',
-          }, {
-            a: 'c',
-          }],
-        },
-      });
-
-      expect(result).toEqual({
-        $set: {
-          avArrayOfObjects: [{
-            a: 'b',
-            foo: 'bar',
-          }, {
-            a: 'c',
-            foo: 'bar',
-          }],
-        },
-      });
-    });
-
-    it('defaults in array of objects', function () {
-      const schema = new SimpleSchema({
-        things: Array,
-        'things.$': Object,
-        'things.$.a': {
-          type: String,
-          defaultValue: 'foo',
-        },
-        'things.$.b': {
-          type: String,
-          defaultValue: 'bar',
-        },
-      });
-
-      const result = schema.clean({
-        $push: {
-          things: {},
-        },
-      });
-
-      expect(result).toEqual({
-        $push: {
-          things: {
-            a: 'foo',
-            b: 'bar',
-          },
-        },
-      });
-    });
-
-    it('$each in autoValue pseudo modifier', function () {
-      const schema = new SimpleSchema({
-        psuedoEach: {
-          type: Array,
-          optional: true,
-          autoValue() {
-            if (this.isSet && this.operator === '$set') {
-              return {
-                $push: {
-                  $each: this.value,
-                },
-              };
-            }
-          },
-        },
-        'psuedoEach.$': {
-          type: String,
-        },
-      });
-
-      const result = schema.clean({
-        $set: {
-          psuedoEach: ['foo', 'bar'],
-        },
-      });
-
-      expect(result).toEqual({
-        $push: {
-          psuedoEach: {
-            $each: ['foo', 'bar'],
-          },
-        },
-      });
-    });
-
-    it('simple autoValues', function () {
-      const schema = new SimpleSchema({
-        name: {
-          type: String,
-        },
-        someDefault: {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.isSet) return 5;
-          },
-        },
-        updateCount: {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.operator) return 0;
-            return { $inc: 1 };
-          },
-        },
-        firstWord: {
-          type: String,
-          optional: true,
-          autoValue() {
-            const content = this.field('content');
-            if (content.isSet) return content.value.split(' ')[0];
-            this.unset();
-          },
-        },
-      });
-
-      let result = schema.clean({
-        name: 'Test',
-        firstWord: 'Illegal to manually set value',
-      });
-
-      expect(result).toEqual({
-        name: 'Test',
-        someDefault: 5,
-        updateCount: 0,
-      });
-
-      result = schema.clean({
-        name: 'Test',
-        someDefault: 20,
-      });
-
-      expect(result).toEqual({
-        name: 'Test',
-        someDefault: 20,
-        updateCount: 0,
-      });
-
-      result = schema.clean({
-        $set: {
-          name: 'Test',
-          someDefault: 20,
-        },
-      });
-
-      expect(result).toEqual({
-        $set: {
-          name: 'Test',
-          someDefault: 20,
-        },
-        $inc: { updateCount: 1 },
-      });
-    });
-
-    it('defaultValue', function () {
-      const defaultValues = new SimpleSchema({
-        name: {
-          type: String,
-          defaultValue: 'Test',
-          optional: true,
-        },
-        a: {
-          type: Object,
-          optional: true,
-        },
-        'a.b': {
-          type: String,
-          defaultValue: 'Test',
-          optional: true,
-        },
-        b: {
-          type: Array,
-          optional: true,
-        },
-        'b.$': {
-          type: Object,
-        },
-        'b.$.a': {
-          type: String,
-          defaultValue: 'Test',
-          optional: true,
-        },
-        strVals: {
-          type: Array,
-          defaultValue: [],
-          optional: true,
-        },
-        'strVals.$': {
-          type: String,
-        },
-      });
-
-      function avClean(obj, exp) {
-        expect(defaultValues.clean(obj)).toEqual(exp);
-      }
-
-      avClean({}, {
-        name: 'Test',
-        a: {
-          b: 'Test',
-        },
-        strVals: [],
-      });
-
-      avClean({
-        strVals: ['foo', 'bar'],
-      }, {
-        name: 'Test',
-        a: {
-          b: 'Test',
-        },
-        strVals: ['foo', 'bar'],
-      });
-
-      avClean({
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-      }, {
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        strVals: [],
-      });
-
-      avClean({
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [],
-      }, {
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [],
-        strVals: [],
-      });
-
-      avClean({
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [{}],
-      }, {
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [{
-          a: 'Test',
-        }],
-        strVals: [],
-      });
-
-      avClean({
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [{
-          a: 'Test1',
-        }, {}],
-      }, {
-        name: 'Test1',
-        a: {
-          b: 'Test1',
-        },
-        b: [{
-          a: 'Test1',
-        }, {
-          a: 'Test',
-        }],
-        strVals: [],
-      });
-
-      // Updates should not be affected, but should get $setOnInsert
-      avClean({
-        $addToSet: {
-          strVals: 'new value',
-        },
-      }, {
-        $addToSet: {
-          strVals: 'new value',
-        },
-        $setOnInsert: {
-          name: 'Test',
-          'a.b': 'Test',
-        },
-      });
-    });
-
-    it('objects in arrays', function () {
-      const subSchema = new SimpleSchema({
-        value: {
-          type: String,
-          autoValue() {
-            expect(this.isSet).toBe(true);
-            expect(this.operator).toEqual('$set');
-            expect(this.value).toEqual('should be overridden by autovalue');
-            return 'autovalue';
-          },
-        },
-      });
-
-      const schema = new SimpleSchema({
-        children: {
-          type: Array,
-        },
-        'children.$': {
-          type: subSchema,
-        },
-      });
-
-      const mod = {
-        $set: {
-          'children.$.value': 'should be overridden by autovalue',
-        },
-      };
-      schema.clean(mod, { mutate: true });
-
-      expect(mod.$set['children.$.value']).toEqual('autovalue');
-    });
-
-    it('operator correct for $pull', function () {
-      let called = false;
-
-      const schema = new SimpleSchema({
-        foo: {
-          type: Array,
-          autoValue() {
-            called = true;
-            expect(this.operator).toEqual('$pull');
-          },
-        },
-        'foo.$': {
-          type: String,
-        },
-      });
-
-      const mod = {
-        $pull: {
-          foo: 'bar',
-        },
-      };
-      schema.clean(mod);
-
-      expect(called).toBe(true);
-    });
-
-    it('issue 340', function () {
-      let called = 0;
-
-      const schema = new SimpleSchema({
-        field1: {
-          type: SimpleSchema.Integer,
-        },
-        field2: {
-          type: String,
-          autoValue() {
-            called++;
-            expect(this.field('field1').value).toBe(1);
-            expect(this.siblingField('field1').value).toBe(1);
-            return 'foo';
-          },
-        },
-      });
-
-      schema.clean({
-        field1: 1,
-      });
-      schema.clean({
-        $set: {
-          field1: 1,
-        },
-      });
-
-      expect(called).toBe(2);
-    });
-
-    it('issue 426', function () {
-      const schema = new SimpleSchema({
-        name: {
-          type: String,
-        },
-        images: {
-          type: Array,
-          label: 'Images',
-          minCount: 0,
-          defaultValue: [],
-        },
-        'images.$': {
-          type: Object,
-          label: 'Image',
-        },
-      });
-
-      const doc = {
-        name: 'Test',
-      };
-      expect(schema.clean(doc)).toEqual({
-        name: 'Test',
-        images: [],
-      });
-    });
-
-    it('clean options should be merged when extending', function () {
-      const schema1 = new SimpleSchema({
-        a: String,
-      }, {
-        clean: {
-          filter: false,
-          autoConvert: false,
-        },
-      });
-
-      const schema2 = new SimpleSchema({});
-      schema2.extend(schema1);
-
-      expect(schema2.clean({ a: 1 })).toEqual({ a: 1 });
-    });
-
-    it('array items', function () {
-      const schema = new SimpleSchema({
-        tags: {
-          type: Array,
-          optional: true,
-        },
-        'tags.$': {
-          type: String,
-          autoValue() {
-            if (this.isSet) return this.value.toLowerCase();
-          },
-        },
-      });
-
-      const obj = {
-        tags: [],
-      };
-      expect(schema.clean(obj)).toEqual({
-        tags: [],
-      });
-
-      const obj2 = {
-        tags: ['FOO', 'BAR'],
-      };
-      expect(schema.clean(obj2)).toEqual({
-        tags: ['foo', 'bar'],
-      });
-    });
-
-    it('updates existing objects when deeply nested (plain)', function () {
-      const schema = new SimpleSchema({
-        nested: Array,
-        'nested.$': Object,
-        'nested.$.doubleNested': Object,
-        'nested.$.doubleNested.integer': {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.value) {
-              return 5;
-            }
-          },
-        },
-      });
-
-      const cleanedObject = schema.clean({
-        nested: [
-          {
-            doubleNested: {},
-          },
-          {
-            doubleNested: {
-              integer: '8',
-            },
-          },
-        ],
-      });
-
-      expect(cleanedObject).toEqual({
-        nested: [
-          {
-            doubleNested: {
-              integer: 5,
-            },
-          },
-          {
-            doubleNested: {
-              integer: 8,
-            },
-          },
-        ],
-      });
-    });
-
-    it('updates existing objects when deeply nested (modifier)', function () {
-      const schema = new SimpleSchema({
-        nested: Array,
-        'nested.$': Object,
-        'nested.$.doubleNested': Object,
-        'nested.$.doubleNested.integer': {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.value) {
-              return 5;
-            }
-          },
-        },
-      });
-
-      const cleanedObject = schema.clean({
-        $push: {
-          nested: {
-            doubleNested: {},
-          },
-        },
-      });
-
-      expect(cleanedObject).toEqual({
-        $push: {
-          nested: {
-            doubleNested: {
-              integer: 5,
-            },
-          },
-        },
-      });
-    });
-
-    it('updates deeply nested with empty $set', function () {
-      const schema = new SimpleSchema({
-        nested: Array,
-        'nested.$': Object,
-        'nested.$.doubleNested': Object,
-        'nested.$.doubleNested.integer': {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.value) {
-              return 5;
-            }
-          },
-        },
-      });
-
-      const cleanedObject = schema.clean({
-        $set: {
-          nested: [{}],
-        },
-      });
-
-      expect(cleanedObject).toEqual({
-        $set: {
-          nested: [{
-            doubleNested: {
-              integer: 5,
-            },
-          }],
-        },
-      });
-    });
-
-    it('updates deeply nested with $set having dotted array key', function () {
-      const schema = new SimpleSchema({
-        nested: Array,
-        'nested.$': Object,
-        'nested.$.doubleNested': Object,
-        'nested.$.doubleNested.integer': {
-          type: SimpleSchema.Integer,
-          autoValue() {
-            if (!this.value) {
-              return 5;
-            }
-          },
-        },
-      });
-
-      const cleanedObject = schema.clean({
-        $set: {
-          'nested.0.doubleNested': {},
-        },
-      });
-
-      expect(cleanedObject).toEqual({
-        $set: {
-          'nested.0.doubleNested': {
-            integer: 5,
-          },
-        },
-      });
-    });
-  });
-
   it('should clean sub schemas', function () {
     const doubleNestedSchema = new SimpleSchema({
       integer: SimpleSchema.Integer,
@@ -1460,96 +516,6 @@ describe('clean', function () {
           integer: 1,
         },
       }],
-    });
-  });
-
-  it('should add auto values to sub schemas for plain objects', function () {
-    const doubleNestedSchema = new SimpleSchema({
-      integer: {
-        type: SimpleSchema.Integer,
-        autoValue() {
-          if (!this.value) {
-            return 5;
-          }
-        },
-      },
-    });
-
-    const nestedSchema = new SimpleSchema({
-      doubleNested: doubleNestedSchema,
-    });
-
-    const schema = new SimpleSchema({
-      nested: Array,
-      'nested.$': nestedSchema,
-    });
-
-    const cleanedObject = schema.clean({
-      nested: [
-        {
-          doubleNested: {},
-        },
-        {
-          doubleNested: {
-            integer: '8',
-          },
-        },
-      ],
-    });
-
-    expect(cleanedObject).toEqual({
-      nested: [
-        {
-          doubleNested: {
-            integer: 5,
-          },
-        },
-        {
-          doubleNested: {
-            integer: 8,
-          },
-        },
-      ],
-    });
-  });
-
-  it('should add auto values to sub schemas for modifiers objects', function () {
-    const doubleNestedSchema = new SimpleSchema({
-      integer: {
-        type: SimpleSchema.Integer,
-        autoValue() {
-          if (!this.value) {
-            return 5;
-          }
-        },
-      },
-    });
-
-    const nestedSchema = new SimpleSchema({
-      doubleNested: doubleNestedSchema,
-    });
-
-    const schema = new SimpleSchema({
-      nested: Array,
-      'nested.$': nestedSchema,
-    });
-
-    const cleanedObject = schema.clean({
-      $push: {
-        nested: {
-          doubleNested: {},
-        },
-      },
-    });
-
-    expect(cleanedObject).toEqual({
-      $push: {
-        nested: {
-          doubleNested: {
-            integer: 5,
-          },
-        },
-      },
     });
   });
 });
