+++ lib/clean/defaultValue.tests.js
@@ -1,0 +1,540 @@
+/* eslint-disable func-names, prefer-arrow-callback */
+
+import expect from 'expect';
+import { SimpleSchema } from '../SimpleSchema';
+
+describe('defaultValue', function () {
+  describe('normal objects', function () {
+    it('adds default value for missing top-level simple prop', function () {
+      const schema = new SimpleSchema({
+        name: {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({});
+
+      expect(result).toEqual({ name: 'Test' });
+    });
+
+    it('does not add default value for already set top-level simple prop', function () {
+      const schema = new SimpleSchema({
+        name: {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({ name: 'Other' });
+
+      expect(result).toEqual({ name: 'Other' });
+    });
+
+    it('adds default value for missing top-level array prop', function () {
+      const schema = new SimpleSchema({
+        names: {
+          type: Array,
+          defaultValue: [],
+          optional: true,
+        },
+        'names.$': String,
+      });
+
+      const result = schema.clean({});
+
+      expect(result).toEqual({ names: [] });
+    });
+
+    it('does not add default value for top-level array prop that is already set', function () {
+      const schema = new SimpleSchema({
+        names: {
+          type: Array,
+          defaultValue: [],
+          optional: true,
+        },
+        'names.$': String,
+      });
+
+      const result = schema.clean({ names: ['foo', 'bar'] });
+
+      expect(result).toEqual({ names: ['foo', 'bar'] });
+    });
+
+    it('does not add defaultValue for prop in object, if object is not set', function () {
+      const schema = new SimpleSchema({
+        a: {
+          type: Object,
+          optional: true,
+        },
+        'a.b': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({});
+
+      expect(result).toEqual({});
+    });
+
+    it('adds defaultValue for prop in object, if object is set in object being cleaned', function () {
+      const schema = new SimpleSchema({
+        a: {
+          type: Object,
+          optional: true,
+        },
+        'a.b': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({ a: {} });
+
+      expect(result).toEqual({ a: { b: 'Test' } });
+    });
+
+    it('adds defaultValue for prop in objects within array, if object is set in object being cleaned', function () {
+      const schema = new SimpleSchema({
+        b: {
+          type: Array,
+          optional: true,
+        },
+        'b.$': {
+          type: Object,
+        },
+        'b.$.a': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({ b: [{}, {}] });
+
+      expect(result).toEqual({ b: [{ a: 'Test' }, { a: 'Test' }] });
+    });
+
+    it('does not add defaultValue for prop in objects within array, if the prop is already set in object being cleaned', function () {
+      const schema = new SimpleSchema({
+        b: {
+          type: Array,
+          optional: true,
+        },
+        'b.$': {
+          type: Object,
+        },
+        'b.$.a': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({ b: [{ a: 'Other' }] });
+
+      expect(result).toEqual({ b: [{ a: 'Other' }] });
+    });
+
+    it('adds defaultValue for prop in objects within array, if array and object are set by array defaultValue', function () {
+      const schema = new SimpleSchema({
+        b: {
+          type: Array,
+          optional: true,
+          defaultValue: [{}, { a: 'Other' }],
+        },
+        'b.$': {
+          type: Object,
+        },
+        'b.$.a': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({});
+
+      expect(result).toEqual({ b: [{ a: 'Test' }, { a: 'Other' }] });
+    });
+
+    it('adds defaultValue for prop in object, if object is set by another defaultValue', function () {
+      const schema = new SimpleSchema({
+        a: {
+          type: Object,
+          defaultValue: {},
+          optional: true,
+        },
+        'a.b': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({});
+
+      expect(result).toEqual({ a: { b: 'Test' } });
+    });
+
+    it('does not add defaultValue for prop in object, if prop already has value in object being cleaned', function () {
+      const schema = new SimpleSchema({
+        a: {
+          type: Object,
+          optional: true,
+        },
+        'a.b': {
+          type: String,
+          defaultValue: 'Test',
+          optional: true,
+        },
+      });
+
+      const result = schema.clean({ a: { b: 'Other' } });
+
+      expect(result).toEqual({ a: { b: 'Other' } });
+    });
+  });
+
+  describe('modifier object', function () {
+    it('adds to $set object', function () {
+      const schema = new SimpleSchema({
+        obj: {
+          type: Object,
+        },
+        'obj.a': {
+          type: Number,
+          optional: true,
+        },
+        'obj.b': {
+          type: Number,
+          optional: true,
+          defaultValue: 10,
+        },
+      });
+
+      const result = schema.clean({
+        $set: {
+          obj: { a: 1 },
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          obj: { a: 1, b: 10 },
+        },
+      });
+    });
+
+    it('adds to $set object with dotted set prop', function () {
+      const schema = new SimpleSchema({
+        obj: {
+          type: Object,
+        },
+        'obj.a': {
+          type: Object,
+          optional: true,
+        },
+        'obj.a.foo': {
+          type: Number,
+          optional: true,
+          defaultValue: 20,
+        },
+        'obj.b': {
+          type: Number,
+          optional: true,
+          defaultValue: 10,
+        },
+      });
+
+      const result = schema.clean({
+        $set: {
+          'obj.a': {},
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a': { foo: 20 },
+        },
+        $setOnInsert: {
+          'obj.b': 10,
+        },
+      });
+    });
+
+    it('adds to $set object with dotted set prop and array', function () {
+      const schema = new SimpleSchema({
+        obj: {
+          type: Object,
+        },
+        'obj.a': {
+          type: Object,
+          optional: true,
+        },
+        'obj.a.foo': {
+          type: Array,
+          optional: true,
+        },
+        'obj.a.foo.$': Object,
+        'obj.a.foo.$.bar': {
+          type: Number,
+          optional: true,
+          defaultValue: 200,
+        },
+      });
+
+      let result = schema.clean({
+        $set: {
+          'obj.a': {},
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a': {},
+        },
+      });
+
+      result = schema.clean({
+        $set: {
+          'obj.a': { foo: [] },
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a': { foo: [] },
+        },
+      });
+
+      result = schema.clean({
+        $set: {
+          'obj.a': { foo: [{}, {}] },
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a': { foo: [{ bar: 200 }, { bar: 200 }] },
+        },
+      });
+    });
+
+    it('adds a $setOnInsert if $setting a sibling prop', function () {
+      const schema = new SimpleSchema({
+        obj: {
+          type: Object,
+        },
+        'obj.a': {
+          type: Number,
+          optional: true,
+        },
+        'obj.b': {
+          type: Number,
+          optional: true,
+          defaultValue: 10,
+        },
+        'obj.c': {
+          type: Number,
+          optional: true,
+          defaultValue: 50,
+        },
+      });
+
+      const result = schema.clean({
+        $set: {
+          'obj.a': 100,
+          'obj.c': 2,
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a': 100,
+          'obj.c': 2,
+        },
+        $setOnInsert: {
+          'obj.b': 10,
+        },
+      });
+    });
+
+    it('adds a $setOnInsert if $setting a sibling child prop', function () {
+      const schema = new SimpleSchema({
+        obj: {
+          type: Object,
+        },
+        'obj.a': {
+          type: Object,
+          optional: true,
+        },
+        'obj.a.one': {
+          type: Number,
+          optional: true,
+          defaultValue: 500,
+        },
+        'obj.a.two': {
+          type: Number,
+          optional: true,
+          defaultValue: 1000,
+        },
+        'obj.b': {
+          type: Number,
+          optional: true,
+          defaultValue: 10,
+        },
+        'obj.c': {
+          type: Number,
+          optional: true,
+          defaultValue: 50,
+        },
+      });
+
+      const result = schema.clean({
+        $set: {
+          'obj.a.one': 100,
+        },
+      });
+
+      expect(result).toEqual({
+        $set: {
+          'obj.a.one': 100,
+        },
+        $setOnInsert: {
+          'obj.a.two': 1000,
+          'obj.b': 10,
+          'obj.c': 50,
+        },
+      });
+    });
+
+    it('adds $setOnInsert for top-level prop', function () {
+      const schema = new SimpleSchema({
+        foo: {
+          type: String,
+          defaultValue: 'Test',
+        },
+        names: {
+          type: Array,
+          optional: true,
+        },
+        'names.$': {
+          type: String,
+        },
+      });
+
+      const result = schema.clean({
+        $addToSet: {
+          names: 'new value',
+        },
+      });
+
+      expect(result).toEqual({
+        $addToSet: {
+          names: 'new value',
+        },
+        $setOnInsert: {
+          foo: 'Test',
+        },
+      });
+    });
+
+    it('adds default values to object being $pushed into array', function () {
+      const schema = new SimpleSchema({
+        things: Array,
+        'things.$': Object,
+        'things.$.a': {
+          type: String,
+          defaultValue: 'foo',
+        },
+        'things.$.b': {
+          type: String,
+          defaultValue: 'bar',
+        },
+      });
+
+      const result = schema.clean({
+        $push: {
+          things: {},
+        },
+      });
+
+      expect(result).toEqual({
+        $push: {
+          things: {
+            a: 'foo',
+            b: 'bar',
+          },
+        },
+      });
+    });
+  });
+
+  it('nested defaultValue for prop in obj in array', function () {
+    const listItemSchema = new SimpleSchema({
+      foo: {
+        type: String,
+        defaultValue: 'TEST',
+        optional: true,
+      },
+    });
+
+    const detailsSchema = new SimpleSchema({
+      list: {
+        type: Array,
+        optional: true,
+      },
+      'list.$': listItemSchema,
+    });
+
+    const schema = new SimpleSchema({
+      name: String,
+      details: {
+        type: detailsSchema,
+        optional: true,
+      },
+    });
+
+    const cleanedObject = schema.clean({
+      name: 'NAME',
+      details: {},
+    });
+
+    expect(cleanedObject).toEqual({
+      name: 'NAME',
+      details: {},
+    });
+  });
+
+  it('issue 426', function () {
+    const schema = new SimpleSchema({
+      name: {
+        type: String,
+      },
+      images: {
+        type: Array,
+        label: 'Images',
+        minCount: 0,
+        defaultValue: [],
+      },
+      'images.$': {
+        type: Object,
+        label: 'Image',
+      },
+    });
+
+    const doc = {
+      name: 'Test',
+    };
+    expect(schema.clean(doc)).toEqual({
+      name: 'Test',
+      images: [],
+    });
+  });
+});
