+++ lib/doValidation.js
@@ -5,7 +5,7 @@ import isObject from 'lodash.isobject';
 import union from 'lodash.union';
 import includes from 'lodash.includes';
 import { SimpleSchema } from './SimpleSchema';
-import { appendAffectedKey, getParentOfKey, looksLikeModifier, isObjectWeShouldTraverse } from './utility.js';
+import { appendAffectedKey, getParentOfKey, looksLikeModifier, isObjectWeShouldTraverse } from './utility';
 import typeValidator from './validation/typeValidator';
 import requiredValidator from './validation/requiredValidator';
 import allowedValuesValidator from './validation/allowedValuesValidator';
@@ -63,7 +63,8 @@ function doValidation({
     }
 
     // Prepare the context object for the validator functions
-    const fieldParentName = getParentOfKey(affectedKey, true);
+    const fieldParentNameWithEndDot = getParentOfKey(affectedKey, true);
+    const fieldParentName = fieldParentNameWithEndDot.slice(0, -1);
 
     function getFieldInfo(key) {
       // Create mongoObject if necessary, cache for speed
@@ -96,8 +97,11 @@ function doValidation({
       key: affectedKey,
       obj,
       operator: op,
+      parentField() {
+        return getFieldInfo(fieldParentName);
+      },
       siblingField(fName) {
-        return getFieldInfo(fieldParentName + fName);
+        return getFieldInfo(fieldParentNameWithEndDot + fName);
       },
       validationContext,
       value: val,
