+++ lib/utility/index.js
@@ -1,3 +1,5 @@
+import getLastPartOfKey from './getLastPartOfKey';
+
 function appendAffectedKey(affectedKey, key) {
   if (key === '$each') return affectedKey;
   return affectedKey ? `${affectedKey}.${key}` : key;
@@ -70,8 +72,9 @@ function getParentOfKey(key, withEndDot) {
 export {
   appendAffectedKey,
   dateToDateString,
+  forEachKeyAncestor,
+  getLastPartOfKey,
+  getParentOfKey,
   isObjectWeShouldTraverse,
   looksLikeModifier,
-  forEachKeyAncestor,
-  getParentOfKey,
 };
