+++ lib/validation/doDateChecks.js
@@ -1,5 +1,5 @@
 import { SimpleSchema } from '../SimpleSchema';
-import { dateToDateString } from '../utility.js';
+import { dateToDateString } from '../utility';
 
 function doDateChecks(def, keyValue) {
   // Is it an invalid date?
